challenge_02/requirements.txt
dj-rest-auth==5.0.1
Django==4.2.6
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pytz==2023.3.post1
requests==2.31.0
setuptools==70.0.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.0.7
uvicorn==0.30.1
pillow==10.4.0
